Appears to work correctly, aside from a vector stack overflow that's difficult to isolate, and the TextOutput box overflows.  Need to fix it by reducing how much content gets printed out to the box.

+++ Utils/custommath.h
@@ -5,7 +5,8 @@
 #include <QtMath>
 #include <QDebug>
 
-typedef long double real;
+//typedef long double real;
+typedef double real;
 
 /**
  * @brief The CustomMath namespace is implemented to store static methods for faster math functions.
